Update README.md
Added tech in the everyday life

My Name is Jason Williams just to introduce who I am at the inception of this idea of a way to change my everyday life and other every day life with an automated way of having better odds of enjoying thier time by saving what it cost. I've studied at a local community college and did complete my courses in Computer and Network Technologies and Science. I'm in no way a degree holder, but I am an enthusiast at least, and some one with a little over 8 years of experience in the technology field and some one who took thier first step in being educated in Technology in the early 2000's. 

I felt like presenting a bit of my wisdom of life into this choice and idea and I do hope its worth as much to you as it is to me as a project. If you wnat to build on this in its beginning stages pre - BETA or pre pre Beta it would be ok with me. 

Im just taking some time to publicly document what im doing but regardless I hope the time invested yields something more than just what I think of it and honestly Im think well of it.

As of October 30th 2024 I have a physical copy of my idea Ill shorten the perspectives and Ideas and if people feel to request a better Idea of the goal of this project you can Email me @ CreatorsHands358@gmail.com or @ JasonCwilliams@protonmail.com. 

Im building this with me and an associates No Profit in mind and the Mission we Have to give back to the community and carry on the good qualities we've been taught in our house hold over the years.

Thank you for your interest in this Project and God Bless:
-------------------------------------------------------------------------------------------------------------------------------------------------------
P.S.

As of : 10/30/2024 23:38

At this state Ive routed my saved information to one single page thats seperate from the intial catalog of information I was given on broiler plate. Ive got a long way to go with my vision of this app but I feel alot more comfortable after some months and time of learning .Net and being educated in and gaining experience in Microprocesses and Large Networks. I found a zen moment of Generalizing operations in C# and .Net. Ive gained a wealth of comprehension and feel to make this project about me using my past experience and relativity of its commonalities to the disipline of Software Creation and Engineering in general. But routing data is what i know and have been taught on the Macro scale and I feel I have a chance to get to becoming an expert in time on the Domain scale. I hope I get this done in a timely fashion and Ill update this node with a simplified version of my entire plan and program. Thank you. 

And when we see a problem we fix a problem..... What I've learned from my time at AT&T.

Just dont become the problem... Lmao. Hope you enjoy.

-------------------------------------------------------------------------------------------------------------------------------------------------------
-------------------------------------------------------------------------------------------------------------------------------------------------------

Books im currently reading:

C#12 in a Nutshell the definitive reference
Pro ASP.NET Core 7 
Clean Architecture with .NET

And Im doing my best to research as much with the Added AI thats available but not in a way to handicap my knowledge by it.

-------------------------------------------------------------------------------------------------------------------------------------------------------
-------------------------------------------------------------------------------------------------------------------------------------------------------

Programming Languages that my be added in a version to this Project

C# 
.NET Framework

(Additions)
SQL
Angular
HTML
JS
CSS
Python

(Needed)
(SWQL)
SolarWinds

(Lite Versions - During BETA)
PostGres
NoSQL
Next.Js

-------------------------------------------------------------------------------------------------------------------------------------------------------
-------------------------------------------------------------------------------------------------------------------------------------------------------